src/requirements.txt

flask>=2.3.3
requests>=2.31.0
azure-storage-blob>=12.17.0
xlsxwriter>=3.1.0
python-dotenv>=1.0.0
